Unify routing table management logic
