+++ ddht/v5_1/network.py
@@ -17,11 +17,11 @@ from eth_enr import ENRAPI, ENRManagerAPI, QueryableENRDatabaseAPI
 from eth_enr.exceptions import OldSequenceNumber
 from eth_typing import NodeID
 from eth_utils import ValidationError
-from eth_utils.toolz import cons, first, take
+from eth_utils.toolz import take
 from lru import LRU
 import trio
 
-from ddht._utils import adaptive_timeout, every, reduce_enrs, weighted_choice
+from ddht._utils import adaptive_timeout, every, reduce_enrs
 from ddht.base_message import InboundMessage
 from ddht.constants import ROUTING_TABLE_BUCKET_SIZE
 from ddht.endpoint import Endpoint
@@ -32,11 +32,9 @@ from ddht.exceptions import (
 )
 from ddht.kademlia import (
     KademliaRoutingTable,
-    at_log_distance,
     compute_log_distance,
     iter_closest_nodes,
 )
-from ddht.token_bucket import TokenBucket
 from ddht.v5_1.abc import (
     ClientAPI,
     DispatcherAPI,
@@ -46,7 +44,6 @@ from ddht.v5_1.abc import (
     PoolAPI,
     TalkProtocolAPI,
 )
-from ddht.v5_1.constants import ROUTING_TABLE_KEEP_ALIVE
 from ddht.v5_1.exceptions import ProtocolNotSupported
 from ddht.v5_1.messages import (
     FindNodeMessage,
@@ -369,9 +366,6 @@ class Network(Service, NetworkAPI):
         self._routing_table_ready.set()
         return True
 
-    async def _bond(self, node_id: NodeID, endpoint: Optional[Endpoint] = None) -> None:
-        await self.bond(node_id, endpoint=endpoint)
-
     async def ping(
         self,
         node_id: NodeID,
@@ -524,107 +518,6 @@ class Network(Service, NetworkAPI):
             self.logger.debug(
                 "routing-table-info: size=%d  buckets=%s", total_size, bucket_info,
             )
-
-    async def _ping_oldest_routing_table_entry(self) -> None:
-        await self._routing_table_ready.wait()
-
-        while self.manager.is_running:
-            # Here we preserve the lazy iteration while still checking that the
-            # iterable is not empty before passing it into `min` below which
-            # throws an ambiguous `ValueError` otherwise if the iterable is
-            # empty.
-            nodes_iter = self.routing_table.iter_all_random()
-            try:
-                first_node_id = first(nodes_iter)
-            except StopIteration:
-                await trio.sleep(ROUTING_TABLE_KEEP_ALIVE)
-                continue
-            else:
-                least_recently_ponged_node_id = min(
-                    cons(first_node_id, nodes_iter),
-                    key=lambda node_id: self._last_pong_at.get(node_id, 0),
-                )
-
-            too_old_at = trio.current_time() - ROUTING_TABLE_KEEP_ALIVE
-            try:
-                last_pong_at = self._last_pong_at[least_recently_ponged_node_id]
-            except KeyError:
-                pass
-            else:
-                if last_pong_at > too_old_at:
-                    await trio.sleep(last_pong_at - too_old_at)
-                    continue
-
-            did_bond = await self.bond(least_recently_ponged_node_id)
-            if not did_bond:
-                self.routing_table.remove(least_recently_ponged_node_id)
-
-    async def _track_last_pong(self) -> None:
-        async with self.dispatcher.subscribe(PongMessage) as subscription:
-            async for message in subscription:
-                self._last_pong_at[message.sender_node_id] = trio.current_time()
-
-    async def _manage_routing_table(self) -> None:
-        # First load all the bootnode ENRs into our database
-        for enr in self._bootnodes:
-            try:
-                self.enr_db.set_enr(enr)
-            except OldSequenceNumber:
-                pass
-
-        # Now repeatedly try to bond with each bootnode until one succeeds.
-        while self.manager.is_running:
-            with trio.move_on_after(20):
-                async with trio.open_nursery() as nursery:
-                    for enr in self._bootnodes:
-                        if enr.node_id == self.local_node_id:
-                            continue
-                        endpoint = Endpoint.from_enr(enr)
-                        nursery.start_soon(self._bond, enr.node_id, endpoint)
-
-                    await self._routing_table_ready.wait()
-                    break
-
-        # Now we enter into an infinite loop that continually probes the
-        # network to beep the routing table fresh.  We both perform completely
-        # random lookups, as well as targeted lookups on the outermost routing
-        # table buckets which are not full.
-        #
-        # The `TokenBucket` allows us to burst at the beginning, making quick
-        # successive probes, then slowing down once the
-        #
-        # TokenBucket starts with 10 tokens, refilling at 1 token every 30
-        # seconds.
-        token_bucket = TokenBucket(1 / 30, 10)
-
-        async with trio.open_nursery() as nursery:
-            while self.manager.is_running:
-                await token_bucket.take()
-
-                # Get the logarithmic distance to the "largest" buckets
-                # that are not full.
-                non_full_bucket_distances = tuple(
-                    idx + 1
-                    for idx, bucket in enumerate(self.routing_table.buckets)
-                    if len(bucket) < self.routing_table.bucket_size  # noqa: E501
-                )[-16:]
-
-                # Probe one of the not-full-buckets with a weighted preference
-                # towards the largest buckets.
-                distance_to_probe = weighted_choice(non_full_bucket_distances)
-                target_node_id = at_log_distance(self.local_node_id, distance_to_probe)
-
-                async with self.recursive_find_nodes(target_node_id) as enr_aiter:
-                    async for enr in enr_aiter:
-                        if enr.node_id == self.local_node_id:
-                            continue
-
-                        try:
-                            self.enr_db.set_enr(enr)
-                        except OldSequenceNumber:
-                            pass
-
-                        nursery.start_soon(self._bond, enr.node_id)
 
     async def _pong_when_pinged(self) -> None:
         async def _maybe_add_to_routing_table(
